+++ root/usr/local/opnsense/mvc/app/controllers/OPNsense/IPsecWatchdog/IndexController.php
@@ -12,6 +12,7 @@ class IndexController extends \OPNsense\Base\IndexController
     {
         $this->view->formDialogTunnel = $this->getForm('tunnel');
         $this->view->formGridTunnel = $this->getFormGrid('tunnel');
+        $this->view->formGeneralSettings = $this->getForm('general');
         $this->view->pick('OPNsense/IPsecWatchdog/index');
     }
 }